% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stl}
\alias{stl}
\title{STL decomposition}
\usage{
stl(.data, colname, window = NULL, ...)
}
\arguments{
\item{.data}{Data frame to decompose.}

\item{colname}{Column name of time-dependent response variable.}

\item{window}{Span of the Loess window.}

\item{...}{
  Arguments passed on to \code{\link[stats:stl]{stats::stl}}
  \describe{
    \item{\code{x}}{univariate time series to be decomposed.
    This should be an object of class \code{"ts"} with a frequency
    greater than one.}
    \item{\code{s.window}}{either the character string \code{"periodic"} or the span (in
    lags) of the loess window for seasonal extraction, which should
    be odd and at least 7, according to Cleveland et al.  This has no default.}
    \item{\code{s.degree}}{degree of locally-fitted polynomial in seasonal
    extraction.  Should be zero or one.}
    \item{\code{t.window}}{the span (in lags) of the loess window for trend
    extraction, which should be odd.  If \code{NULL}, the default,
    \code{nextodd(ceiling((1.5*period) / (1-(1.5/s.window))))}, is taken.}
    \item{\code{t.degree}}{degree of locally-fitted polynomial in trend
    extraction.  Should be zero or one.}
    \item{\code{l.window}}{the span (in lags) of the loess window of the low-pass
    filter used for each subseries.  Defaults to the smallest odd
    integer greater than or equal to \code{frequency(x)} which is
    recommended since it prevents competition between the trend and
    seasonal components.  If not an odd integer its given value is
    increased to the next odd one.}
    \item{\code{l.degree}}{degree of locally-fitted polynomial for the subseries
    low-pass filter.  Must be 0 or 1.}
    \item{\code{s.jump}}{integers at least one to increase speed of
    the respective smoother.  Linear interpolation happens between every
    \code{*.jump}th value.}
    \item{\code{t.jump}}{integers at least one to increase speed of
    the respective smoother.  Linear interpolation happens between every
    \code{*.jump}th value.}
    \item{\code{l.jump}}{integers at least one to increase speed of
    the respective smoother.  Linear interpolation happens between every
    \code{*.jump}th value.}
    \item{\code{robust}}{logical indicating if robust fitting be used in the
    \code{loess} procedure.}
    \item{\code{inner}}{integer; the number of \sQuote{inner} (backfitting)
    iterations; usually very few (2) iterations suffice.}
    \item{\code{outer}}{integer; the number of \sQuote{outer} robustness
    iterations.}
    \item{\code{na.action}}{action on missing values.}
  }}
}
\value{
Model data frame, a \code{\link[fabletools]{mable}}, with \code{stl} model
column.
}
\description{
Changes the default \code{robust} from \code{FALSE} to \code{TRUE}; estimates
a robust STL by default.
}
\examples{
\dontshow{if (requireNamespace("tsibble", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tsibble::as_tsibble(USAccDeaths) \%>\%
  isaiahr::stl("value") \%>\%
  fabletools::components() \%>\%
  fabletools::autoplot()
\dontshow{\}) # examplesIf}
}
