% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{stl_components}
\alias{stl_components}
\title{STL decomposition components}
\usage{
stl_components(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=stl]{stl}}
  \describe{
    \item{\code{.data}}{Data frame to decompose.}
    \item{\code{colname}}{Column name of time-dependent response variable.}
    \item{\code{window}}{Span of the Loess window.}
  }}
}
\value{
Decomposition table (\code{\link[fabletools]{dable}}) comprising
trend, season and remainder.
}
\description{
STL decomposition components
}
\examples{
\dontshow{if (requireNamespace("tsibble", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tsibble::as_tsibble(USAccDeaths) \%>\%
  isaiahr::stl_components("value") \%>\%
  fabletools::autoplot()
\dontshow{\}) # examplesIf}
}
