% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_retail_employment}
\alias{us_retail_employment}
\title{Filtered employment data by Retail Trade}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 357 rows and 3 columns.
}
\usage{
us_retail_employment
}
\description{
Filtered employment data by Retail Trade
}
\examples{
library(fabletools)
library(ggplot2)

autoplot(us_retail_employment, Employed) +
  labs(y = "Persons (thousands)",
       title = "Total employment in US retail")
}
\keyword{datasets}
